[add] MDX plugins [optimize] update Upstream packages
