+++ package.json
@@ -14,13 +14,13 @@
     "file-type": "^18.5.0",
     "html2canvas": "^1.4.1",
     "idea-react": "^1.0.0-rc.22",
-    "koajax": "^0.9.2",
+    "koajax": "^0.9.4",
     "leaflet": "^1.9.4",
     "leaflet.chinatmsproviders": "^3.0.6",
     "less": "^4.2.0",
     "less-loader": "^11.1.3",
     "license-filter": "^0.2.3",
-    "marked": "^8.0.0",
+    "marked": "^8.0.1",
     "mobx": "^6.10.2",
     "mobx-i18n": "^0.4.1",
     "mobx-lark": "^1.0.0-rc.7",
@@ -43,17 +43,24 @@
     "webpack": "^5.88.2"
   },
   "devDependencies": {
+    "@mdx-js/loader": "^2.3.0",
+    "@mdx-js/react": "^2.3.0",
+    "@next/mdx": "^12.3.4",
     "@types/leaflet": "^1.9.4",
-    "@types/node": "^18.17.14",
+    "@types/mdx": "^2.0.7",
+    "@types/node": "^18.17.15",
     "@types/react": "^17.0.65",
     "@types/react-dom": "^17.0.20",
-    "eslint": "^8.48.0",
+    "eslint": "^8.49.0",
     "eslint-config-next": "^12.3.4",
     "eslint-plugin-simple-import-sort": "^10.0.0",
     "get-git-folder": "^0.1.2",
     "husky": "^8.0.3",
     "lint-staged": "^14.0.1",
     "prettier": "^3.0.3",
+    "remark-frontmatter": "^4.0.1",
+    "remark-gfm": "^3.0.1",
+    "remark-mdx-frontmatter": "^3.0.0",
     "typescript": "~5.1.6"
   },
   "pnpm": {
